Move disjoint-set into importer directory.
